requirements.txt
flask
numpy
PyYAML
requests
pillow
matplotlib
mysql.connector
twilio
